SonicNet/requirements.txt
matplotlib=3.6.0=pypi_0
numpy=1.23.3=pypi_0
opencv-python=4.6.0.66=pypi_0
pyqt6=6.4.0=pypi_0
pyqt6-qt6=6.4.1=pypi_0
pyqt6-sip=13.4.0=pypi_0
pyserial=3.5=pypi_0
pyyaml=6.0=pypi_0
torch=1.12.1=pypi_0
torchvision=0.13.1=pypi_0
